refactor: englishCard
