+++ src/Widget/EnglishCard/EnglishCard.tsx
@@ -4,12 +4,12 @@ import styled from 'styled-components';
 import { Swiper, SwiperSlide } from 'swiper/react';
 import ReactLoading from 'react-loading';
 
+import Controller from './Controller';
 import { getCard } from './fireStore';
 import EnglishWord, { EnglishWordData } from './EnglishWord';
 import { handleNewEnglishWords, afterOneDay } from './utils';
-import { DoubleArrow, Refresh } from '../../components/Icons';
-import { SwiperButtonNext, SwiperButtonPrev } from '../../components/Swiper';
-import useLocalStorage from '../../hooks/useLocalStorage';
+
+import { useLocalStorage } from '../../hooks';
 
 import 'swiper/css';
 
@@ -21,61 +21,6 @@ const Wrapper = styled.div`
   & > .swiper {
     width: 100%;
     height: 100%;
-  }
-`;
-
-const IconsContainer = styled.div`
-  position: absolute;
-  bottom: 10px;
-  right: 10px;
-  z-index: 5;
-  display: flex;
-  gap: 5px;
-
-  & div:not(:last-child) {
-    visibility: hidden;
-  }
-
-  & svg {
-    fill: ${({ theme }) => theme.color.transparentWhite};
-  }
-  :hover {
-    & div:not(:last-child) {
-      visibility: visible;
-    }
-
-    & svg {
-      fill: ${({ theme }) => theme.color.lightWhite};
-    }
-  }
-`;
-
-const IconWrapper = styled.div`
-  padding: 2px;
-  width: 28px;
-  height: 28px;
-  border-radius: 50%;
-  cursor: pointer;
-
-  & > button {
-    background: transparent;
-    border: none;
-    padding: 0;
-    width: 100%;
-    height: 100%;
-    cursor: pointer;
-  }
-
-  &:hover {
-    background: ${({ theme }) => theme.color.transparentWhite};
-  }
-  &:hover svg {
-    fill: ${({ theme }) => theme.color.white};
-  }
-
-  & svg {
-    width: 24px;
-    height: 24px;
   }
 `;
 
@@ -155,21 +100,7 @@ function EnglishCard({ data }: EnglishCardProps) {
             <EnglishWord word={word} tags={tag} />
           </SwiperSlide>
         ))}
-        <IconsContainer>
-          <IconWrapper>
-            <SwiperButtonPrev>
-              <DoubleArrow direction="left" />
-            </SwiperButtonPrev>
-          </IconWrapper>
-          <IconWrapper onClick={updateWords}>
-            <Refresh />
-          </IconWrapper>
-          <IconWrapper>
-            <SwiperButtonNext>
-              <DoubleArrow direction="right" />
-            </SwiperButtonNext>
-          </IconWrapper>
-        </IconsContainer>
+        <Controller updateWords={updateWords} />
       </Swiper>
       {isLoading && (
         <LoadingWrapper>
